Refactor: python, cts: Move NearQuorumPointTest into the _cts module.
No code changes.
